Removed dispatcher and using simple calls
