+++ plugins/ptp_operator/socket/socket_test.go
@@ -2,44 +2,39 @@ package socket_test
 
 import (
 	"bufio"
-	ptp_event "github.com/redhat-cne/cloud-event-proxy/plugins/ptp_operator/event"
+	"github.com/redhat-cne/cloud-event-proxy/plugins/ptp_operator/metrics"
 	ptp_socket "github.com/redhat-cne/cloud-event-proxy/plugins/ptp_operator/socket"
 	log "github.com/sirupsen/logrus"
 	"github.com/stretchr/testify/assert"
 	"net"
-	"sync"
 	"testing"
 	"time"
 )
 
-var jobDispatcher *ptp_event.Dispatcher
-var dispatchWG sync.WaitGroup
-const logLength =15
-var logsData=[logLength]string{
-	"ptp4l[3535499.401]: [ens5f1] port 1: delay timeout" + "\n",
-	"ptp4l[3535499.402]: [ens5f1] delay   filtered         88   raw         82" + "\n",
-	"ptp4l[3535499.424]: [ens5f1] master offset          1 s2 freq   -1869 path delay        88" + "\n",
-	"ptp4l[3535499.476]: [ens5f1] port 1: delay timeout" + "\n",
-	"ptp4l[3535499.476]: [ens5f1] delay   filtered         88   raw         87" + "\n",
-	"ptp4l[3535499.485]: [ens5f1] port 1: delay timeout" + "\n",
-	"ptp4l[3535499.485]: [ens5f1] delay   filtered         88   raw         88" + "\n",
-	"ptp4l[3535499.488]: [ens5f1] master offset         12 s2 freq   -1850 path delay        88" + "\n",
-	"ptp4l[3535432.615]: [ens5f1] port 1: UNCALIBRATED to SLAVE on MASTER_CLOCK_SELECTED" + "\n",
-	"phc2sys[3535433.762]: [ens5f1] reconfiguring after port state change" + "\n",
-	"phc2sys[3535433.762]: [ens5f1] selecting CLOCK_REALTIME for synchronization" + "\n",
-	"phc2sys[3535433.762]: [ens5f1] selecting ens5f0 as the master" + "\n",
-	"phc2sys[96254.969]: [ens5f1] CLOCK_REALTIME phc offset      100 s2 freq  -79243 delay   1058" + "\n",
-	"phc2sys[432313.127]: [ens5f1] CLOCK_REALTIME phc offset   -837364 s2 freq +625227 delay   1415",
-	"ptp4l[432313.222]: [ens5f1] port 1: SLAVE to FAULTY on FAULT_DETECTED (FT_UNSPECIFIED)" +"\n",
-}
 
-func Setup() {
-	jobDispatcher = ptp_event.NewDispatcher(&dispatchWG)
-	jobDispatcher.Start()
+const logLength = 15
 
+var logsData = [logLength]string{
+	"ptp4l[3535499.401]: [ens5f1] port 1: delay timeout " + "\n",
+	"ptp4l[3535499.402]: [ens5f1] delay   filtered         88   raw         82 " + "\n",
+	"ptp4l[3535432.615]: [ens5f1] port 1: UNCALIBRATED to SLAVE on MASTER_CLOCK_SELECTED " + "\n",
+	"ptp4l[3535499.424]: [ens5f1] master offset          1 s0 freq   -1869 path delay        88 " + "\n",
+	"ptp4l[3535499.476]: [ens5f1] port 1: delay timeout " + "\n",
+	"ptp4l[3535499.476]: [ens5f1] delay   filtered         88   raw         87 " + "\n",
+	"ptp4l[3535499.485]: [ens5f1] port 1: delay timeout " + "\n",
+	"ptp4l[3535499.485]: [ens5f1] delay   filtered         88   raw         88 " + "\n",
+	"ptp4l[3535499.488]: [ens5f1] master offset         12 s2 freq   -1850 path delay        88 " + "\n",
+	"phc2sys[3535433.762]: [ens5f1] reconfiguring after port state change " + "\n",
+	"phc2sys[3535433.762]: [ens5f1] selecting CLOCK_REALTIME for synchronization " + "\n",
+	"phc2sys[3535433.762]: [ens5f1] selecting ens5f0 as the master " + "\n",
+	"phc2sys[96254.969]: [ens5f1] CLOCK_REALTIME phc offset      100 s2 freq  -79243 delay   1058 " + "\n",
+	"phc2sys[432313.127]: [ens5f1] CLOCK_REALTIME phc offset   -837364 s2 freq +625227 delay   1415 " + "\n",
+	"ptp4l[432313.222]: [ens5f1] port 1: SLAVE to FAULTY on FAULT_DETECTED (FT_UNSPECIFIED) " + "\n",
 }
+var metricsProcessor *metrics.Metric
+
 func Test_WriteMetricsToSocket(t *testing.T) {
-	Setup()
+	metricsProcessor = &metrics.Metric{Stats: make(map[string]*metrics.Stats)}
 	go listenToTestMetrics()
 	time.Sleep(2 * time.Second)
 	c, err := net.Dial("unix", "/tmp/go.sock")
@@ -50,19 +45,14 @@ func Test_WriteMetricsToSocket(t *testing.T) {
 
 	defer c.Close()
 
-	for i:=0;i<logLength;i++{
+	for i := 0; i < logLength; i++ {
 		_, err = c.Write([]byte(logsData[i]))
 		if err != nil {
 			log.Fatal("write error:", err)
 		}
 	}
-	go func() {
-		time.Sleep(5 * time.Second)
-		jobDispatcher.Stop()
-	}()
 
-	dispatchWG.Wait()
-	//time.Sleep(2*time.Second)
+	time.Sleep(3*time.Second)
 }
 
 func listenToTestMetrics() {
@@ -96,7 +86,8 @@ func processTestMetrics2(c net.Conn) {
 		}
 		log.Printf("plugin got %s", scanner.Text())
 		msg := scanner.Text()
-		jobDispatcher.ProcessMsg(msg)
+		metricsProcessor.ExtractMetrics(msg)
 	}
 
 }
+
